Move Spell Icon stuff to Old namespace. Prep for rewrite
