+++ EndlessClient/HUD/Panels/Old/OldActiveSpellsPanel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using EndlessClient.Dialogs;
-using EndlessClient.HUD.Spells;
+using EndlessClient.HUD.Spells.Old;
 using EndlessClient.Old;
 using EndlessClient.UIControls;
 using EOLib;
